@@ -1,17 +1,37 @@
-import React, { useState, useEffect } from 'react';
+import { BrowserRouter, Navigate, Routes, Route } from "react-router-dom";
+import HomePage from "scenes/homePage/homePage";
+import LogInPage from "scenes/loginPage/LogInPage";
+import ProfilePage from "scenes/profilePage/ProfilePage";
+import { useSelector } from "react-redux";
+import { CssBaseline, ThemeProvider } from "@mui/material";
+import { createTheme } from "@mui/material/styles";
+import { themeSettings } from "./theme";
+import { useMemo } from "react";
+
 
 function App() {
-  const [backendData, setBackendData] = useState("");
-
-  useEffect(() => {
-    fetch('/')
-      .then(response => response.json())
-      .then(data => setBackendData(data.message));
-  }, []);
+  const mode = useSelector((state) => state.mode);
+  const theme = useMemo(() => createTheme(themeSettings(mode)), [mode]);
+  const isAuth = Boolean(useSelector((state) => state.token));
 
   return (
-    <div>
-      <h1>Hello from team 9 also known as MizFits!</h1>
+    <div className="app">
+      <BrowserRouter>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          <Routes>
+            <Route path="/" element={<LogInPage />} />
+            <Route
+              path="/home"
+              element={isAuth ? <HomePage /> : <Navigate to="/" />}
+            />
+            <Route
+              path="/profile/:userId"
+              element={isAuth ? <ProfilePage /> : <Navigate to="/" />}
+            />
+          </Routes>
+        </ThemeProvider>
+      </BrowserRouter>
     </div>
   );
 }
